requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
openai==1.40.0
azure-search-documents==11.4.0
azure-identity==1.16.0
azure-cosmos==4.7.0
pydantic==2.7.0
pydantic-settings==2.3.0
structlog==24.2.0
python-dotenv==1.0.1
httpx==0.27.0
python-multipart==0.0.9
pytest==8.2.0
pytest-asyncio==0.23.0
